requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
line-bot-sdk==3.21.0
pyngrok==7.4.1
requests==2.32.5
psycopg2-binary==2.9.9